Solved the logout cookie issue.
Issue: The logout page successfully destroyed the session and cookies, but under that the header.php was included which (being still on the server had the cookies) so generated a new session. Then had to reload to delete that session, and then the logout was over. [Solved]

Mostly the sign in up out is done. Also fixed sessions on all pages.